refactor: 뉴스 상세 컨텐츠 리팩터링
